--- packages/ciphernode/enclave/src/commands/mod.rs
@@ -1,68 +1,0 @@
-pub mod aggregator;
-pub mod init;
-pub mod net;
-pub mod password;
-pub mod start;
-pub mod wallet;
-
-use self::password::PasswordCommands;
-use aggregator::AggregatorCommands;
-use clap::Subcommand;
-use net::NetCommands;
-use wallet::WalletCommands;
-
-#[derive(Subcommand, Debug)]
-pub enum Commands {
-    /// Start the application
-    Start,
-
-    /// Aggregator node management commands
-    Aggregator {
-        #[command(subcommand)]
-        command: AggregatorCommands,
-    },
-
-    /// Password management commands
-    Password {
-        #[command(subcommand)]
-        command: PasswordCommands,
-    },
-
-    /// Wallet management commands
-    Wallet {
-        #[command(subcommand)]
-        command: WalletCommands,
-    },
-
-    /// Networking related commands
-    Net {
-        #[command(subcommand)]
-        command: NetCommands,
-    },
-
-    Init {
-        /// An rpc url for enclave to connect to
-        #[arg(long = "rpc-url")]
-        rpc_url: Option<String>,
-
-        /// An Ethereum address that enclave should use to identify the node
-        #[arg(long = "eth-address")]
-        eth_address: Option<String>,
-
-        /// The password
-        #[arg(short, long)]
-        password: Option<String>,
-
-        /// Skip asking for eth
-        #[arg(long = "skip-eth")]
-        skip_eth: bool,
-
-        /// The network private key (ed25519)
-        #[arg(long = "net-keypair")]
-        net_keypair: Option<String>,
-
-        /// Generate a new network keypair
-        #[arg(long = "generate-net-keypair")]
-        generate_net_keypair: bool,
-    },
-}
