+++ packages/ciphernode/enclave/src/helpers/prompt_password.rs
