+++ packages/ciphernode/enclave/src/net.rs
@@ -1,9 +1,8 @@
-mod generate;
-mod purge;
-mod set;
 use anyhow::*;
 use clap::Subcommand;
 use config::AppConfig;
+
+use crate::{net_generate, net_purge, net_set};
 
 #[derive(Subcommand, Debug)]
 pub enum NetCommands {
@@ -22,9 +21,9 @@ pub enum NetCommands {
 
 pub async fn execute(command: NetCommands, config: &AppConfig) -> Result<()> {
     match command {
-        NetCommands::PurgeId => purge::execute(&config).await?,
-        NetCommands::GenerateKey => generate::execute(&config).await?,
-        NetCommands::SetKey { net_keypair } => set::execute(&config, net_keypair).await?,
+        NetCommands::PurgeId => net_purge::execute(&config).await?,
+        NetCommands::GenerateKey => net_generate::execute(&config).await?,
+        NetCommands::SetKey { net_keypair } => net_set::execute(&config, net_keypair).await?,
     };
 
     Ok(())
