+++ packages/ciphernode/enclave/src/net_generate.rs
