+++ packages/ciphernode/enclave/src/password_overwrite.rs
@@ -1,5 +1,5 @@
-use super::create::execute as set_password;
-use super::delete::{prompt_delete, DeleteMode};
+use super::password_create::execute as set_password;
+use super::password_delete::{prompt_delete, DeleteMode};
 use anyhow::Result;
 use config::AppConfig;
